llmspell_hooks::traits

Trait Hook

pub trait Hook: Send + Sync {
    // Required methods
    fn execute<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: &'life1 mut HookContext,
    ) -> Pin<Box<dyn Future<Output = Result<HookResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn metadata(&self) -> HookMetadata { ... }
    fn should_execute(&self, _context: &HookContext) -> bool { ... }
}

Base hook trait for all hooks in the system

Required Methods

fn execute<'life0, 'life1, 'async_trait>( &'life0 self, context: &'life1 mut HookContext, ) -> Pin<Box<dyn Future<Output = Result<HookResult>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Execute the hook with the given context

fn as_any(&self) -> &dyn Any

Get self as Any for downcasting

Provided Methods

fn metadata(&self) -> HookMetadata

Get metadata about this hook

fn should_execute(&self, _context: &HookContext) -> bool

Check if this hook should be executed for the given context